button refactor for use in quesiton list
