+++ src/Components/ConfirmModal/ConfirmModal.tsx
@@ -1,4 +1,4 @@
-import React, { ReactNode, useState } from "react";
+import React, { ReactNode } from "react";
 import Button from "react-bootstrap/Button";
 import Modal from "react-bootstrap/Modal";
