+++ src/Components/QuestionList/QuestionCards/QuestionCardsS.scss
@@ -5,10 +5,6 @@
 }
 
 .QuestionCard {
-    .listItem.card {
-        display: flex;
-        flex-direction: column;
-        justify-content: flex-start;
-        width: 20rem;
-    }
+    display: flex;
+    flex-direction: row;
 }
