+++ src/Components/QuestionList/QuestionList.tsx
@@ -18,12 +18,14 @@ import {
     Tags,
     filterSettingsDefault,
 } from "../../Apis/types";
+import { Button } from "react-bootstrap";
 
 export const QuestionList = () => {
     const { questions } = useContext(MyQuestionContext);
     const [filterSettings, setFilterSettings] = useState<FilterSetting>(
         filterSettingsDefault
     );
+    const [showDelete, setShowDelete] = useState<boolean>(false);
 
     const { name, type, hideCorrect, difficulty } = filterSettings;
 
@@ -49,11 +51,22 @@ export const QuestionList = () => {
     return (
         <div className="QuestionList">
             <AddQuestionForm />
+            <Button
+                onClick={() => {
+                    setShowDelete(!showDelete);
+                }}
+            >
+                Edit Questions
+            </Button>
             <div className="filter-container">
                 <QuestionFilters filterSettings={setFilterSettings} />
             </div>
 
-            <QuestionCards currentQuestions={currentQuestions} />
+            <QuestionCards
+                currentQuestions={currentQuestions}
+                setShowDelete={setShowDelete}
+                showDelete={showDelete}
+            />
         </div>
     );
 };
