+++ src/storage.json
@@ -19,22 +19,6 @@
     "questions": [
         {
             "difficulty": "medium",
-            "question": "what method executes a provided function once for each array element.",
-            "answer": "Array.forEach()",
-            "answers": [
-                "Array.map()",
-                "Array.forEach()",
-                "Array.every()"
-            ],
-            "tags": [
-                "array",
-                "Js"
-            ],
-            "correct": false,
-            "id": "3"
-        },
-        {
-            "difficulty": "medium",
             "question": " What method tests whether at least one element in the array passes the test implemented by the provided function. It returns true if, in the array, it finds an element for which the provided function returns true; otherwise it returns false. It doesn't modify the array.",
             "answer": "Array.some()",
             "answers": [
@@ -64,22 +48,6 @@
             ],
             "correct": false,
             "id": "5"
-        },
-        {
-            "difficulty": "hard",
-            "question": "this is a test obj to see if the filter is working.",
-            "answer": "test1",
-            "answers": [
-                "test1",
-                "test",
-                "test"
-            ],
-            "tags": [
-                "object",
-                "Js"
-            ],
-            "correct": true,
-            "id": "6"
         },
         {
             "difficulty": "medium",
